# vulkan: apply MUL_MAT_ID subgroup optimization to non-coopmat devices