This applies the optimization @jeffbolznv contributed in #15427 to non-coopmat GPUs, as long as they support subgroup ballots. It closes the gap on AMD GPUs to ROCm prompt processing in Mixture-of-Experts models.

ggml_vulkan: 0 = AMD Radeon (TM) Pro VII (RADV VEGA20) (radv) | uma: 0 | fp16: 1 | bf16: 0 | warp size: 64 | shared memory: 65536 | int dot: 1 | matrix cores: none
| model                          | size      | params  | backend | ngl | fa | test  | t/s (before)  | t/s (after)   | diff    |
|--------------------------------|-----------|---------|---------|-----|----|-------|---------------|---------------|---------|
| qwen3moe 30B.A3B Q2_K - Medium | 10.48 GiB | 30.53 B | Vulkan  | 99  | 0  | pp512 | 45.16 ± 0.06  | 342.07 ± 3.44 | +657.5% |
| qwen3moe 30B.A3B Q2_K - Medium | 10.48 GiB | 30.53 B | Vulkan  | 99  | 1  | pp512 | 44.83 ± 0.02  | 317.92 ± 3.07 | +609.2% |
| gpt-oss 20B Q8_0               | 11.27 GiB | 20.91 B | Vulkan  | 99  | 0  | pp512 | 191.17 ± 0.68 | 520.07 ± 5.21 | +172.0% |
| gpt-oss 20B Q8_0               | 11.27 GiB | 20.91 B | Vulkan  | 99  | 1  | pp512 | 190.88 ± 0.65 | 508.50 ± 5.39 | +166.4% |

ggml_vulkan: 0 = AMD Radeon RX 6800 XT (RADV NAVI21) (radv) | uma: 0 | fp16: 1 | bf16: 0 | warp size: 32 | shared memory: 65536 | int dot: 1 | matrix cores: none
| model                          | size      | params  | backend | ngl | fa | test  | t/s (before)  | t/s (after)     | diff    |
|--------------------------------|-----------|---------|---------|-----|----|-------|---------------|-----------------|---------|
| qwen3moe 30B.A3B Q2_K - Medium | 10.48 GiB | 30.53 B | Vulkan  | 99  | 0  | pp512 | 137.88 ± 0.09 | 781.56 ± 4.59   | +466.8% |
| qwen3moe 30B.A3B Q2_K - Medium | 10.48 GiB | 30.53 B | Vulkan  | 99  | 1  | pp512 | 136.86 ± 0.08 | 758.64 ± 3.46   | +454.3% |
| gpt-oss 20B Q8_0               | 11.27 GiB | 20.91 B | Vulkan  | 99  | 0  | pp512 | 584.38 ± 1.23 | 1181.19 ± 6.38  | +102.1% |
| gpt-oss 20B Q8_0               | 11.27 GiB | 20.91 B | Vulkan  | 99  | 1  | pp512 | 590.68 ± 2.70 | 1200.36 ± 13.60 | +103.2% |

ggml_vulkan: 0 = Intel(R) Arc(tm) A770 Graphics (DG2) (Intel open-source Mesa driver) | uma: 0 | fp16: 1 | bf16: 1 | warp size: 32 | shared memory: 65536 | int dot: 1 | matrix cores: none
| model                          | size      | params  | backend | ngl | fa | test  | t/s (before)  | t/s (after)   | diff    |
|--------------------------------|-----------|---------|---------|-----|----|-------|---------------|---------------|---------|
| qwen3moe 30B.A3B Q2_K - Medium | 10.48 GiB | 30.53 B | Vulkan  | 99  | 0  | pp512 | 35.25 ± 0.04  | 127.06 ± 0.70 | +260.5% |
| qwen3moe 30B.A3B Q2_K - Medium | 10.48 GiB | 30.53 B | Vulkan  | 99  | 1  | pp512 | 30.81 ± 0.04  | 82.30 ± 0.30  | +167.1% |
| gpt-oss 20B Q8_0               | 11.27 GiB | 20.91 B | Vulkan  | 99  | 0  | pp512 | 115.28 ± 0.74 | 177.98 ± 1.80 | +54.4%  |
| gpt-oss 20B Q8_0               | 11.27 GiB | 20.91 B | Vulkan  | 99  | 1  | pp512 | 114.72 ± 0.49 | 175.38 ± 1.30 | +52.9%  |


ggml_vulkan: 0 = NVIDIA GeForce RTX 3090 (NVIDIA) | uma: 0 | fp16: 1 | bf16: 1 | warp size: 32 | shared memory: 49152 | int dot: 1 | matrix cores: none
| model                          | size      | params  | backend | ngl | fa | test  | t/s (before)  | t/s (after)    | diff    |
|--------------------------------|-----------|---------|---------|-----|----|-------|---------------|----------------|---------|
| qwen3moe 30B.A3B Q2_K - Medium | 10.48 GiB | 30.53 B | Vulkan  | 99  | 0  | pp512 | 458.25 ± 2.08 | 960.08 ± 3.09  | +109.5% |
| qwen3moe 30B.A3B Q2_K - Medium | 10.48 GiB | 30.53 B | Vulkan  | 99  | 1  | pp512 | 453.05 ± 1.17 | 949.62 ± 7.72  | +109.6% |
| gpt-oss 20B Q8_0               | 11.27 GiB | 20.91 B | Vulkan  | 99  | 0  | pp512 | 914.31 ± 5.93 | 1224.32 ± 9.70 | +33.9%  |
| gpt-oss 20B Q8_0               | 11.27 GiB | 20.91 B | Vulkan  | 99  | 1  | pp512 | 914.47 ± 8.57 | 1217.99 ± 7.75 | +33.2%  |

(Nvidia result without coopmat or coopmat2, those already have the optimization)

@netrunnereve This should give your gpt-oss benchmarks a good boost.