@netrunnereve 2025-08-24
@3Simplex Since this is happening on master as well you should create a new issue as it's not really relavant to this PR. FYI in the future please use llama-bench to get your results.

Your issue might be due to vram and you can try monitoring that while you run with different context sizes.