@3Simplex 2025-08-24
> @0cc4m just FYI, since I assumed you made these tables manually: take a look at `scipts/compare-commits.sh`. The script in turn internally calls `llama-bench` and `scripts/compare-commits.py` to create a table with a performance comparison automatically.

This is exactly the kind of thing I needed to know.

> @3Simplex Since this is happening on master as well you should create a new issue as it's not really relavant to this PR. FYI in the future please use llama-bench to get your results.
> 
> Your issue might be due to vram and you can try monitoring that while you run with different context sizes.

Yeah, that's what made me laugh above. I have been meaning to learn about other functionality like the scripts and examples but haven't had the time to read them. I was thinking, that would have made things so much easier if I had known about it earlier.