@MrLavender 2025-08-25
I see no difference in performance on my RX 6800 16GB (not XT) between b6262 and b6264 (which contains this PR). I'm using RADV 25.1.7 on Linux (also tried AMDVLK 2.0.349). How do I investigate this?

I'm using the  `llama-bXXXX-bin-ubuntu-vulkan-x64.zip` assets from Releases for the Vulkan tests.

b6262
```
$ llama-bench -fa 0,1 -m gpt-oss-20b-mxfp4.gguf
load_backend: loaded RPC backend from /home/xxx/.local/llama-cpp/bin/libggml-rpc.so
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = AMD Radeon RX 6800 (RADV NAVI21) (radv) | uma: 0 | fp16: 1 | bf16: 0 | warp size: 32 | shared memory: 65536 | int dot: 1 | matrix cores: none
load_backend: loaded Vulkan backend from /home/xxx/.local/llama-cpp/bin/libggml-vulkan.so
load_backend: loaded CPU backend from /home/xxx/.local/llama-cpp/bin/libggml-cpu-haswell.so
| model                          |       size |     params | backend    | ngl | fa |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | -: | --------------: | -------------------: |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | RPC,Vulkan |  99 |  0 |           pp512 |        482.25 ± 1.55 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | RPC,Vulkan |  99 |  0 |           tg128 |        125.94 ± 0.06 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | RPC,Vulkan |  99 |  1 |           pp512 |        485.12 ± 1.67 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | RPC,Vulkan |  99 |  1 |           tg128 |        123.76 ± 0.14 |

build: c9a24fb9 (6262)
```

b6264
```
$ llama-bench -fa 0,1 -m gpt-oss-20b-mxfp4.gguf
load_backend: loaded RPC backend from /home/xxx/.local/llama-cpp/bin/libggml-rpc.so
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = AMD Radeon RX 6800 (RADV NAVI21) (radv) | uma: 0 | fp16: 1 | bf16: 0 | warp size: 32 | shared memory: 65536 | int dot: 1 | matrix cores: none
load_backend: loaded Vulkan backend from /home/xxx/.local/llama-cpp/bin/libggml-vulkan.so
load_backend: loaded CPU backend from /home/xxx/.local/llama-cpp/bin/libggml-cpu-haswell.so
| model                          |       size |     params | backend    | ngl | fa |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | -: | --------------: | -------------------: |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | RPC,Vulkan |  99 |  0 |           pp512 |        481.72 ± 1.60 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | RPC,Vulkan |  99 |  0 |           tg128 |        125.78 ± 0.10 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | RPC,Vulkan |  99 |  1 |           pp512 |        485.46 ± 1.55 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | RPC,Vulkan |  99 |  1 |           tg128 |        123.61 ± 0.09 |

build: 043fb27d (6264)
```
ROCm
```
ggml_cuda_init: GGML_CUDA_FORCE_MMQ:    no
ggml_cuda_init: GGML_CUDA_FORCE_CUBLAS: no
ggml_cuda_init: found 1 ROCm devices:
  Device 0: AMD Radeon RX 6800, gfx1030 (0x1030), VMM: no, Wave Size: 32
| model                          |       size |     params | backend    | ngl | fa |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | -: | --------------: | -------------------: |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | ROCm       |  99 |  0 |           pp512 |      1711.88 ± 12.16 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | ROCm       |  99 |  0 |           tg128 |         95.92 ± 0.11 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | ROCm       |  99 |  1 |           pp512 |       1254.50 ± 5.84 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | ROCm       |  99 |  1 |           tg128 |         93.35 ± 0.35 |

build: c9a24fb9 (6262)
```
Vulkan subgroup info if it helps;

```
$ vulkaninfo | grep group
        VK_KHR_device_group_creation           : extension revision 1
        subgroupSize                      = 64
        subgroupSupportedStages: count = 14
        subgroupSupportedOperations: count = 10
        subgroupQuadOperationsInAllStages = true
        minSubgroupSize                                                               = 32
        maxSubgroupSize                                                               = 64
        maxComputeWorkgroupSubgroups                                                  = 4294967295
        requiredSubgroupSizeStages: count = 3
        VK_EXT_pipeline_library_group_handles         : extension revision 1
        VK_EXT_shader_subgroup_ballot                 : extension revision 1
        VK_EXT_shader_subgroup_vote                   : extension revision 1
        VK_EXT_subgroup_size_control                  : extension revision 2
        VK_KHR_device_group                           : extension revision 4
        VK_KHR_shader_subgroup_extended_types         : extension revision 1
        VK_KHR_shader_subgroup_rotate                 : extension revision 2
        VK_KHR_shader_subgroup_uniform_control_flow   : extension revision 1
        VK_KHR_workgroup_memory_explicit_layout       : extension revision 1
        VK_KHR_zero_initialize_workgroup_memory       : extension revision 1
        shaderSubgroupClock = true
VkPhysicalDeviceShaderSubgroupUniformControlFlowFeaturesKHR:
        shaderSubgroupUniformControlFlow = true
        shaderSubgroupExtendedTypes                        = true
        subgroupBroadcastDynamicId                         = true
        subgroupSizeControl                                = true
        computeFullSubgroups                               = true
        shaderZeroInitializeWorkgroupMemory                = true
        shaderSubgroupRotate                   = true
        shaderSubgroupRotateClustered          = true
VkPhysicalDeviceWorkgroupMemoryExplicitLayoutFeaturesKHR:
        workgroupMemoryExplicitLayout                  = true
        workgroupMemoryExplicitLayoutScalarBlockLayout = true
        workgroupMemoryExplicitLayout8BitAccess        = true
        workgroupMemoryExplicitLayout16BitAccess       = true
```